pyproject.toml
[tool.poetry]
name = "django-queued-storage"
version = "0.9.0"
description = "Queued remote storage for Django."
readme = "README.rst"
authors = [
    "Jannis Leidel <jannis@leidel.info>",
    "Sam Bishop <sam@techdragon.io>"
]
license = "BSD"
classifiers=[
        'Development Status :: 4 - Beta',
        'Framework :: Django',
        'Intended Audience :: Developers',
        'License :: OSI Approved :: BSD License',
        'Operating System :: OS Independent',
        'Programming Language :: Python',
        'Programming Language :: Python :: 2.7',
        'Programming Language :: Python :: 3.3',
        'Programming Language :: Python :: 3.4',
        'Topic :: Utilities',
]
packages = [
    { include = "queued_storage" },
]

[tool.poetry.urls]
homepage = 'https://github.com/jazzband/django-queued-storage'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
django_find_project = false
#addopts = '--cov queued_storage'


[tool.poetry.dependencies]
python = "^3.8"
six = ">=1.10.0"
django-appconf = ">=0.4"
packaging = ">=16.8"
setuptools-scm = "^8.0.4"
celery = "^5.3.6"
django = "3.2.16"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-sugar = "^0.9.7"
pytest-assume = "^2.4.3"
pytest-celery = "^0.0.0"
pytest-django = "^4.7.0"
django-cleanup = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
